Added iteration and build tools
The iterRow and iterColumn methods created iterable objects that can traverse the matrix row by row, or column by column.

The buildRow method returns an instance of the buildMatrix class. Instances of this class can be used to build a matrix entry by entry using the insert method. Once a matrix is built, it can be converted into an instance of the matrix class using the get method.